minimax and new evaluation function work
